Fix from Jos for batch mode on Solaris.

+++ batch.c
@@ -396,7 +396,7 @@ void write_batch_csum_info(int *flist_entry, int flist_count,
 			   struct sum_struct *s)
 {
 	size_t i;
-	unsigned int int_zero = 0;
+	size_t int_zero = 0;
 	extern int csum_length;
 
 	fdb_open = 1;
